requirements.txt
pygame>=2.5
python-chess>=1.999
pytest>=8.0
